requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
spacy>=3.5.0
scikit-image>=0.19.0
pytesseract>=0.3.10
pyperclip>=1.8.0
joblib>=1.2.0
matplotlib>=3.5.0
pillow>=9.0.0
scikit-learn>=1.0.0